Move comparison tests out of event_test.py since that module is out of control.
